Search query - Search per date query - Search per list query - Search per sign (>=, <=, [interval]) Bug : search per date bug fixed

+++ app/src/main/java/com/openclassrooms/realestatemanager/models/SearchListItem.kt
@@ -1,7 +1,7 @@
 package com.openclassrooms.realestatemanager.models
 
 class SearchListItem : ISearchListItem {
-    override var id: Int = 0
+    override var id: Long = 0
     override var title: String? = null
 
 }
